+++ knockoff/factory/sink.py
@@ -4,10 +4,10 @@
 # This source code is licensed under the Apache-2.0 license found in
 # the LICENSE file in the root directory of this source tree.
 
-from knockoff.utilities.mixin import FactoryMixin
+from knockoff.utilities.mixin import ResourceLocatorMixin
 
 
-class SinkDumpStrategyFactory(FactoryMixin, object):
+class SinkDumpStrategyFactory(ResourceLocatorMixin, object):
     entry_point_group = "knockoff.factory.sink.dump_strategy"