+++ knockoff/sdk/factory/column.py
@@ -9,6 +9,15 @@ import numpy as np
 from faker import Faker
 
 
+class ColumnFactory(object):
+    def __init__(self, column, callable_):
+        self.column = column
+        self.callable = callable_
+
+    def __call__(self, *args, **kwargs):
+        return {self.column: self.callable(*args,**kwargs)}
+
+
 class ChoiceFactory(object):
     def __init__(self, choices, p=None, replace=True):
         self.choices = choices