+++ knockoff/sdk/table.py
@@ -17,6 +17,9 @@ from faker import Faker
 from knockoff.exceptions import FactoryNotFound, AttemptLimitReached
 
 
+from .factory.column import ColumnFactory
+
+
 logger = logging.getLogger(__name__)
 
 
@@ -170,9 +173,17 @@ class KnockoffTable(object):
         """TODO: should this be public?"""
         data = {}
         for factory in self.factories:
+            key_values = {}
             if isinstance(factory, (tuple, list)):
+                logger.warning("Using a tuple will be "
+                               "deprecated in newer releases."
+                               "Use ColumnFactory instead")
                 col, factory = factory
                 key_values = {col: factory()}
+            elif (isinstance(factory, (ColumnFactory,)) and
+                   factory.depends_on):
+                kwargs = {key: data[key] for key in factory.depends_on}
+                key_values = factory(**kwargs)
             else:
                 key_values = factory()
             data.update(key_values)